requirements.txt
requests>=2.31.0
googlesearch-python>=1.2.3
shodan>=1.31.0
pyfiglet>=0.8.0
colorama>=0.4.6